Day4/Gen app4/requirements.txt
load_dotenv
langchain
langchain_ollama
langchain_huggingface
langchain_community
